ldap_publisher/requirements.txt
credstash
boto3 #no-deploy
botocore #no-deploy
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.22.0
cffi==1.10.0
configobj==5.0.6
cryptography==2.0
docutils==0.13.1
everett==0.9
idna==2.5
jmespath==0.9.2
jsonschema==2.6.0
packaging==16.8
pathtools==0.1.2
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
watchdog==0.8.3
pluginbase
pykmssig
git+git://github.com/mozilla-iam/cis.git@85947b70b21f1ad9d4878519d79b127d2e5fcd56#egg=mozilla_iam_cis
